adaptix/services/hrms/apps/noticeboard/serializers.py
from rest_framework import serializers
from .models import Notice

class NoticeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notice
        fields = '__all__'
        read_only_fields = ('company_uuid', 'created_by', 'created_at', 'updated_at')
